Add some guns, the inventory images
