+++ vitrix/assets/models/pistol.mtl
@@ -1,19 +1,8 @@
 # Blender MTL File: 'None'
-# Material Count: 2
-
-newmtl Material.001
-Ns 323.999994
-Ka 1.000000 1.000000 1.000000
-Kd 0.800000 0.800000 0.800000
-Ks 0.500000 0.500000 0.500000
-Ke 0.000000 0.000000 0.000000
-Ni 1.450000
-d 1.000000
-illum 2
-refl pistol_UVmap_Paint.png
+# Material Count: 1
 
 newmtl None
-Ns 500.000001
+Ns 499.999983
 Ka 1.000000 1.000000 1.000000
 Kd 0.800000 0.800000 0.800000
 Ks 0.800000 0.800000 0.800000
